Added proxy agent to be called from the outside
